gltf28: Auto-add actions to the NLA editor
The Khronos glTF exporter uses the NLA editor to control export of
actions. If a given object has no NLA tracks, but has actions, they are
now auto-added to the NLA editor so they get exported. If a user wants
to control which actions are exported for an object, they can still use
the NLA editor as normal.

This replicates the behavior of the 2.7x glTF pipeline where all actions
are automatically exported.