cli: Force all materials to be single-sided by default
By default, new Blender materials are double-sided, which can be a big
performance impact if not disabled. Since most users do not want
double-sided materials and this is an easy setting to miss, we will
force all materials to be single-sided by default. To disable this and
have blend2bam respect the Blender material settings, use the new
'--allow-double-sided-materials' flag.